added react routes and refactored the code to fit this
