+++ friends/src/components/FriendList.js
@@ -1,44 +1,14 @@
-import React, { Component } from 'react';
-import axios from 'axios';
+import React from 'react';
 
 import Friend from './Friend';
 
-class FriendList extends Component {
-    constructor() {
-        super();
-        this.state = {
-            friends: []
-        };
-    }
+const FriendList = props =>  (
+    <div>
+        {props.friends.map(friend => {
+            return <Friend key={friend[0]} friend={friend} />
+        })}
+    </div>
+);
 
-    componentDidMount() {
-        const self = this;
-        axios.get('http://localhost:5000/friends')
-            .then(function (response) {
-
-                const friendArr = [];
-                response.data.map(friend => {
-                    friendArr.push([friend.id, friend.name, friend.age, friend.email])
-                });
-
-                self.setState({
-                    friends: friendArr
-                });
-            })
-            .catch(function (error) {
-                console.log(error);
-            });
-    }
-
-    render() {
-        return (
-            <div>
-                {this.state.friends.map(friend => {
-                    return <Friend key={friend[0]} friend={friend} />
-                })}
-            </div>
-        );
-    }
-}
 
 export default FriendList;
